# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Healer Hospital Management System API",
    "description": "Comprehensive API for managing hospital operations including appointments, staff, departments, and patient care",
    "version": "1.0.0",
    "contact": {
      "name": "Healer Development Team",
      "email": "dev@healer.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.healer.vn/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["BOOKED", "PAID", "CANCEL"],
            "example": "BOOKED"
          },
          "patientId": {
            "type": "string"
          },
          "medicalRoomId": {
            "type": "string"
          },
          "bookingTimeId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Shift": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "fromTime": {
            "type": "string",
            "format": "date-time"
          },
          "toTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Staff": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstname": {
            "type": "string",
            "example": "Dr. Jane"
          },
          "lastname": {
            "type": "string",
            "example": "Smith"
          },
          "introduction": {
            "type": "string",
            "example": "Experienced cardiologist"
          },
          "educationLevel": {
            "type": "string",
            "enum": ["HIGHSCHOOL", "BACHELOR", "MASTER", "DOCTORATE"]
          },
          "accountId": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/appointments": {
      "get": {
        "summary": "Get appointments",
        "description": "Retrieve a list of appointments",
        "tags": ["Appointments"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["BOOKED", "PAID", "CANCEL"]
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Appointment"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create appointment",
        "description": "Book a new appointment",
        "tags": ["Appointments"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "patientId", "medicalRoomTimeId"],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "patientId": {
                    "type": "string"
                  },
                  "medicalRoomTimeId": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully"
          }
        }
      }
    },
...
    "/shifts": {
      "get": {
        "summary": "Get shifts",
        "description": "Retrieve a list of work shifts",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shifts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Shift"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create shift",
        "description": "Create a new work shift",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["doctorId", "roomId", "fromTime", "toTime"],
                "properties": {
                  "doctorId": {
                    "type": "string"
                  },
                  "roomId": {
                    "type": "string"
                  },
                  "fromTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "toTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shift created successfully"
          }
        }
      }
    },
    "/staff": {
      "get": {
        "summary": "Get staff",
        "description": "Retrieve a list of staff members",
        "tags": ["Staff"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Staff retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Staff"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}